Added new schedule, some test commands, general improvements

+++ src/com/github/xerragnaroek/jikai/anime/alrh/ARHandler.java
@@ -1,23 +1,3 @@
-/*
- * MIT License
- *
- * Copyright (c) 2019 github.com/XerRagnaroek
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
- * associated documentation files (the "Software"), to deal in the Software without restriction,
- * including without limitation the rights to use, copy, modify, merge, publish, distribute,
- * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or
- * substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
- * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
- * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
 package com.github.xerragnaroek.jikai.anime.alrh;
 
 import java.util.Set;
@@ -28,6 +8,7 @@ import org.slf4j.LoggerFactory;
 
 import com.github.xerragnaroek.jasa.Anime;
 import com.github.xerragnaroek.jikai.anime.db.AnimeDB;
+import com.github.xerragnaroek.jikai.anime.db.AnimeUpdate;
 import com.github.xerragnaroek.jikai.jikai.Jikai;
 import com.github.xerragnaroek.jikai.user.JikaiUser;
 import com.github.xerragnaroek.jikai.util.BotUtils;
@@ -126,7 +107,9 @@ class ARHandler {
 					if (mr.getCount() == 1) {
 						log.debug("No user reaction left for {}", title);
 						ALRHData data = alrhDB.getDataForTitle(title);
-						data.setReacted(false);
+						if (data.isReacted()) {
+							data.setReacted(false);
+						}
 					}
 					found = true;
 					break;
@@ -143,23 +126,29 @@ class ARHandler {
 		tc.retrieveMessageById(msgId).submit().whenComplete((m, t) -> {
 			long tmp = msgId;
 			if (t == null) {
-				data.forEach(d -> validateReaction(m, d.getUnicodeCodePoint(), d.getTitle()));
+				if (m != null) {
+					data.forEach(d -> validateReaction(m, d.getUnicodeCodePoint(), d.getTitle()));
+				} else {
+					log.error("Saved message id is invalid!");
+				}
 			} else {
 				log.error("Message not found", t);
 			}
 		});
 	}
 
-	void update() {
-		Set<ALRHData> reacted = alrhDB.getReactedAnimes();
-		Set<String> titles = AnimeDB.getSeasonalAnime().stream().map(Anime::getTitleRomaji).collect(Collectors.toSet());
-		reacted.forEach(data -> {
-			String title = data.getTitle();
-			if (!titles.contains(title)) {
-				alrhDB.deleteEntry(data);
-				log.info("Deleted obsolete Role and data for {}", title);
-			}
-		});
+	void update(AnimeUpdate au) {
+		if (au.hasRemovedAnime()) {
+			Set<ALRHData> reacted = alrhDB.getReactedAnimes();
+			Set<String> titles = au.getRemovedAnime().stream().map(Anime::getTitleRomaji).collect(Collectors.toSet());
+			reacted.forEach(data -> {
+				String title = data.getTitle();
+				if (!titles.contains(title)) {
+					alrhDB.deleteEntry(data);
+					log.info("Deleted obsolete data for {}", title);
+				}
+			});
+		}
 	}
 
 }
